refactor: 에러 코드 enum 정의 및 예외 파싱 유틸 함수 분리
