+++ src/common/constants/error-code.enum.ts
@@ -1,0 +1,4 @@
+export enum ErrorCode {
+  DUPLICATE_ERROR = 'DUPLICATE_ERROR',
+  VALIDATION_ERROR = 'VALIDATION_ERROR',
+}
